ref(issue-stream): Refactor IssueListSearchBar to be more reusable
