+++ static/app/views/issueList/searchBar.tsx
@@ -1,17 +1,14 @@
 import {useCallback} from 'react';
+
 // eslint-disable-next-line no-restricted-imports
-import {withRouter, WithRouterProps} from 'react-router';
-
+import {fetchTagValues} from 'sentry/actionCreators/tags';
 import SmartSearchBar from 'sentry/components/smartSearchBar';
-import {
-  makePinSearchAction,
-  makeSaveSearchAction,
-} from 'sentry/components/smartSearchBar/actions';
-import {SavedSearch, SavedSearchType, Tag, TagCollection} from 'sentry/types';
+import {Organization, SavedSearchType, Tag, TagCollection} from 'sentry/types';
+import {getUtcDateString} from 'sentry/utils/dates';
 import {FieldKind, getFieldDefinition} from 'sentry/utils/fields';
-import useOrganization from 'sentry/utils/useOrganization';
-
-import {TagValueLoader} from './types';
+import useApi from 'sentry/utils/useApi';
+import usePageFilters from 'sentry/utils/usePageFilters';
+import withIssueTags, {WithIssueTagsProps} from 'sentry/utils/withIssueTags';
 
 const getSupportedTags = (supportedTags: TagCollection) =>
   Object.fromEntries(
@@ -26,24 +23,38 @@ const getSupportedTags = (supportedTags: TagCollection) =>
     ])
   );
 
-interface Props extends React.ComponentProps<typeof SmartSearchBar>, WithRouterProps {
-  onSidebarToggle: () => void;
-  sort: string;
-  supportedTags: TagCollection;
-  tagValueLoader: TagValueLoader;
-  savedSearch?: SavedSearch;
+interface Props extends React.ComponentProps<typeof SmartSearchBar>, WithIssueTagsProps {
+  organization: Organization;
 }
 
-function IssueListSearchBar({
-  onSidebarToggle: _onSidebarToggle,
-  sort,
-  supportedTags,
-  tagValueLoader,
-  savedSearch,
-  location,
-  ...props
-}: Props) {
-  const organization = useOrganization();
+const EXCLUDED_TAGS = ['environment'];
+
+function IssueListSearchBar({organization, tags, ...props}: Props) {
+  const api = useApi();
+  const {selection: pageFilters} = usePageFilters();
+
+  const tagValueLoader = useCallback(
+    (key: string, search: string) => {
+      const orgId = organization.slug;
+      const projectIds = pageFilters.projects.map(id => id.toString());
+      const endpointParams = {
+        start: getUtcDateString(pageFilters.datetime.start),
+        end: getUtcDateString(pageFilters.datetime.end),
+        statsPeriod: pageFilters.datetime.period,
+      };
+
+      return fetchTagValues(api, orgId, key, search, projectIds, endpointParams);
+    },
+    [
+      api,
+      organization.slug,
+      pageFilters.datetime.end,
+      pageFilters.datetime.period,
+      pageFilters.datetime.start,
+      pageFilters.projects,
+    ]
+  );
+
   const getTagValues = useCallback(
     async (tag: Tag, query: string): Promise<string[]> => {
       const values = await tagValueLoader(tag.key, query);
@@ -52,28 +63,17 @@ function IssueListSearchBar({
     [tagValueLoader]
   );
 
-  const pinnedSearch = savedSearch?.isPinned ? savedSearch : undefined;
-
   return (
     <SmartSearchBar
-      searchSource="main_search"
       hasRecentSearches
       savedSearchType={SavedSearchType.ISSUE}
       onGetTagValues={getTagValues}
-      actionBarItems={[
-        makePinSearchAction({sort, pinnedSearch, location}),
-        makeSaveSearchAction({
-          sort,
-          disabled: !organization.access.includes('org:write'),
-        }),
-        // TODO: fully remove this
-        // makeSearchBuilderAction({onSidebarToggle}),
-      ]}
-      {...props}
+      excludedTags={EXCLUDED_TAGS}
       maxMenuHeight={500}
-      supportedTags={getSupportedTags(supportedTags)}
+      supportedTags={getSupportedTags(tags)}
+      {...props}
     />
   );
 }
 
-export default withRouter(IssueListSearchBar);
+export default withIssueTags(IssueListSearchBar);
